refactor(sidecar): extract PDProxyManager interface
- Add PDProxyManager interface to decouple runners from concrete manager
- Move manager.go to pd_proxy_manager.go
- Add GetDecoderProxy() method to ProxyManager
- Update go:generate directive to include PDProxyManager in mocks
